requirements.txt
# Azure Functions
azure-functions==1.17.0

# Azure Services
azure-storage-blob==12.19.0
azure-communication-sms==1.0.1
azure-communication-phonenumbers==1.1.0
azure-eventgrid==4.22.0
azure-cognitiveservices-vision-computervision==0.9.0

# Document processing
pypdf==3.17.4
python-docx==1.1.0

# Azure OpenAI
openai==1.3.0

# Redis
redis==5.0.1

# Configuration and validation
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development dependencies (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0 